# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostSummary } from './blogPostSummary';

export type BlogPost = BlogPostSummary & ({
  body?: string | null;
  seoTitle?: string | null;
  seoDescription?: string | null;
});