# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DestinationStatus } from './destinationStatus';

export interface Destination {
  id: string;
  slug: string;
  name: string;
  fullName?: string | null;
  tagline?: string | null;
  description?: string | null;
  heroImage?: string | null;
  images: string[];
  highlights: string[];
  bestTime?: string | null;
  distance?: string | null;
  mapEmbed?: string | null;
  seoTitle?: string | null;
  seoDescription?: string | null;
  status?: DestinationStatus;
  sortOrder?: number;
}