# lib/api-zod/src/generated/types/destinationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DestinationInputStatus } from './destinationInputStatus';

export interface DestinationInput {
  slug: string;
  name: string;
  fullName?: string | null;
  tagline?: string | null;
  description?: string | null;
  heroImage?: string | null;
  images?: string[];
  highlights?: string[];
  bestTime?: string | null;
  distance?: string | null;
  mapEmbed?: string | null;
  seoTitle?: string | null;
  seoDescription?: string | null;
  status?: DestinationInputStatus;
  sortOrder?: number;
}