# lib/api-zod/src/generated/types/galleryItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItemInput {
  src: string;
  alt?: string | null;
  caption?: string | null;
  category: string;
  tall?: boolean;
  sortOrder?: number;
  published?: boolean;
}